FE-1074 - remove cases folder
